+++ package.json
@@ -22,7 +22,8 @@
     "react-router-dom": "^6.14.2",
     "react-switch": "^7.0.0",
     "reactflow": "^11.8.2",
-    "styled-components": "^5.3.11"
+    "styled-components": "^5.3.11",
+    "tailwindcss": "^3.4.17"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
