+++ src/App.css
@@ -1,6 +1,5 @@
 #root {
   text-align: center;
-  background-color: #fff;
   display: flex;
   flex-direction: column;
   min-height: 100vh;
