+++ src/components/content/Block.jsx
@@ -1,15 +1,14 @@
 import styled from "styled-components";
-import { THEME } from "../../utils";
 
 const Root = styled.div`
   display: block;
-  background-color: ${THEME.BKG};
+  background-color: var(--colour-bkg-2);
   padding: 0.5em 2em;
   margin: 1em 0em;
   height: ${(props) => props?.height || "auto"};
   width: ${(props) => props?.width || "auto"};
   border-radius: 8px; /*Optional:Rounded*/
-  box-shadow: 0 4px 6px rgba(0,0,0,0.1); /*Static shadow effect */
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); /*Static shadow effect */
 
   h2 {
     text-align: left;
@@ -17,8 +16,8 @@ const Root = styled.div`
 
   p {
     text-align: left;
-    color: ${THEME.SECACCENT};
     margin-left: 0em;
+    margin-bottom: 2rem;
   }
 `;
