+++ src/utils.js
@@ -2,7 +2,7 @@ export const THEME = {
   ACCENT: "#E2B046",
   PRIMARY: "#345800",
   BKG: "#F0EFE3",
-  SECACCENT: "#224352"
+  SECACCENT: "#224352",
 };
 
 export const PAGES = [
